@@ -1,39 +1,49 @@
 {
     "type": "project",
     "license": "proprietary",
+    "minimum-stability": "stable",
+    "prefer-stable": true,
     "require": {
-        "php": "^8.0.0",
+        "php": ">=8.0",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "composer/package-versions-deprecated": "^1.11",
-        "knplabs/knp-markdown-bundle": "^1.8",
-        "sensio/framework-extra-bundle": "^6.0",
-        "sentry/sentry-symfony": "^4.0",
-        "symfony/asset": "5.1.*",
-        "symfony/console": "5.1.*",
-        "symfony/dotenv": "5.1.*",
+        "composer/package-versions-deprecated": "1.11.99.3",
+        "doctrine/doctrine-bundle": "^2.4",
+        "doctrine/doctrine-migrations-bundle": "^3.1",
+        "doctrine/orm": "^2.9",
+        "knplabs/knp-markdown-bundle": "^1.9",
+        "knplabs/knp-time-bundle": "^1.16",
+        "sensio/framework-extra-bundle": "^6.2",
+        "stof/doctrine-extensions-bundle": "^1.5.0",
+        "symfony/asset": "5.3.*",
+        "symfony/console": "5.3.*",
+        "symfony/debug-bundle": "5.3.*",
+        "symfony/dotenv": "5.3.*",
         "symfony/flex": "^1.3.1",
-        "symfony/framework-bundle": "5.1.*",
-        "symfony/monolog-bundle": "^3.0",
-        "symfony/stopwatch": "5.1.*",
-        "symfony/twig-bundle": "5.1.*",
-        "symfony/webpack-encore-bundle": "^1.7",
-        "symfony/yaml": "5.1.*",
+        "symfony/framework-bundle": "5.3.*",
+        "symfony/monolog-bundle": "^3.7",
+        "symfony/proxy-manager-bridge": "5.3.*",
+        "symfony/routing": "5.3.*",
+        "symfony/runtime": "5.3.*",
+        "symfony/twig-bundle": "5.3.*",
+        "symfony/webpack-encore-bundle": "^1.12",
+        "symfony/yaml": "5.3.*",
         "twig/extra-bundle": "^2.12|^3.0",
         "twig/twig": "^2.12|^3.0"
     },
     "require-dev": {
-        "symfony/debug-bundle": "5.1.*",
-        "symfony/maker-bundle": "^1.15",
-        "symfony/var-dumper": "5.1.*",
-        "symfony/web-profiler-bundle": "5.1.*"
+        "doctrine/doctrine-fixtures-bundle": "^3.4",
+        "symfony/maker-bundle": "^1.33",
+        "symfony/stopwatch": "5.3.*",
+        "symfony/web-profiler-bundle": "5.3.*",
+        "zenstruck/foundry": "^1.13"
     },
     "config": {
+        "optimize-autoloader": true,
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true,
-        "platform": {}
+        "sort-packages": true
     },
     "autoload": {
         "psr-4": {
@@ -46,13 +56,9 @@
         }
     },
     "replace": {
-        "paragonie/random_compat": "2.*",
         "symfony/polyfill-ctype": "*",
         "symfony/polyfill-iconv": "*",
-        "symfony/polyfill-php72": "*",
-        "symfony/polyfill-php71": "*",
-        "symfony/polyfill-php70": "*",
-        "symfony/polyfill-php56": "*"
+        "symfony/polyfill-php72": "*"
     },
     "scripts": {
         "auto-scripts": {
@@ -71,8 +77,8 @@
     },
     "extra": {
         "symfony": {
-            "allow-contrib": false,
-            "require": "5.1.*"
+            "allow-contrib": true,
+            "require": "5.3.*"
         }
     }
 }
