@@ -9,9 +9,7 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class CommentController extends AbstractController
 {
-    /**
-     * @Route("/comments/{id}/vote/{direction<up|down>}", methods="POST")
-     */
+    #[Route('/comments/{id}/vote/{direction<up|down>}', methods: 'POST')]
     public function commentVote($id, $direction, LoggerInterface $logger)
     {
         // todo - use id to query the database
