@@ -2,9 +2,13 @@
 
 namespace App\Controller;
 
+use App\Entity\Question;
+use App\Repository\QuestionRepository;
 use App\Service\MarkdownHelper;
+use Doctrine\ORM\EntityManagerInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -20,18 +24,24 @@ class QuestionController extends AbstractController
     }
 
 
-    /**
-     * @Route("/", name="app_homepage")
-     */
-    public function homepage()
+    #[Route('/', name: 'homepage')]
+    public function homepage(QuestionRepository $repository)
     {
-        return $this->render('question/homepage.html.twig');
+        $questions = $repository->findAllAskedOrderedByNewest();
+        return $this->render('question/homepage.html.twig',[
+            'questions' => $questions,
+        ]);
+    }
+    #[Route('questions/new', name: 'question_new')]
+
+    public function new(EntityManagerInterface $entityManager)
+    {
+        return new Response('Sounds like nice feature for V2');
     }
 
-    /**
-     * @Route("/questions/{slug}", name="app_question_show")
-     */
-    public function show($slug, MarkdownHelper $markdownHelper)
+    #[Route('questions/{slug}', name: 'question_show')]
+
+    public function show(Question $question)
     {
         if ($this->isDebug) {
             $this->logger->info('We are in debug mode!');
@@ -42,14 +52,29 @@ class QuestionController extends AbstractController
             'Honestly, I like furry shoes better than MY cat',
             'Maybe... try saying the spell backwards?',
         ];
-        $questionText = 'I\'ve been turned into a cat, any *thoughts* on how to turn back? While I\'m **adorable**, I don\'t really care for cat food.';
-
-        $parsedQuestionText = $markdownHelper->parse($questionText);
 
         return $this->render('question/show.html.twig', [
-            'question' => ucwords(str_replace('-', ' ', $slug)),
-            'questionText' => $parsedQuestionText,
+            'question' => $question,
             'answers' => $answers,
         ]);
     }
+    #[Route('/questions/{slug}/vote', name: 'question_vote', methods: 'POST')]
+    public function questionVote(Question $question, Request $request, EntityManagerInterface $entityManager)
+    {
+        $direction = $request->request->get('direction');
+        if($direction === "up")
+        {
+            $question->upVote();
+        } elseif ($direction === "down")
+        {
+            $question->downVote();
+        }
+
+        $entityManager->flush();
+
+        return $this->redirectToRoute('question_show',[
+            'slug' => $question->getSlug()
+    ]);
+    }
+
 }
